refactor: move get_validator_info to EpochManagerAdapter (#7727)
part of https://github.com/near/nearcore/issues/6910, depends on https://github.com/near/nearcore/pull/7726
